fix bug

+++ tkbc/process_roland_data.py
@@ -46,7 +46,8 @@ def prepare_dataset(path, name):
     n_entities = len(entities)
 
     os.makedirs(os.path.join(DATA_PATH, name), exist_ok=True)
-    os.remove(os.path.join(DATA_PATH, name, f))
+    if os.path.exists(os.path.join(DATA_PATH, name, f)):
+        os.remove(os.path.join(DATA_PATH, name, f))
     # write ent to id / rel to id
     for (dic, f) in zip([entities_to_id, relations_to_id, timestamps_to_id], ['ent_id', 'rel_id', 'ts_id']):
         ff = open(os.path.join(DATA_PATH, name, f), 'w+')
